+++ src/titiler/application/titiler/application/settings.py
@@ -1,9 +1,10 @@
 """Titiler API settings."""
 
-import pydantic
+from pydantic import field_validator
+from pydantic_settings import BaseSettings, SettingsConfigDict
 
 
-class ApiSettings(pydantic.BaseSettings):
+class ApiSettings(BaseSettings):
     """FASTAPI application settings."""
 
     name: str = "TiTiler"
@@ -18,13 +19,9 @@ class ApiSettings(pydantic.BaseSettings):
 
     lower_case_query_parameters: bool = False
 
-    @pydantic.validator("cors_origins")
+    model_config = SettingsConfigDict(env_prefix="TITILER_API_", env_file=".env")
+
+    @field_validator("cors_origins")
     def parse_cors_origin(cls, v):
         """Parse CORS origins."""
         return [origin.strip() for origin in v.split(",")]
-
-    class Config:
-        """model config"""
-
-        env_file = ".env"
-        env_prefix = "TITILER_API_"