+++ src/titiler/core/titiler/core/dependencies.py
@@ -1,7 +1,6 @@
 """Common dependency."""
 
 import json
-import sys
 from dataclasses import dataclass
 from enum import Enum
 from typing import Dict, List, Optional, Sequence, Tuple, Union
@@ -12,12 +11,7 @@ from rasterio.crs import CRS
 from rio_tiler.colormap import cmap, parse_color
 from rio_tiler.errors import MissingAssets, MissingBands
 from rio_tiler.types import ColorMapType, RIOResampling
-
-if sys.version_info >= (3, 9):
-    from typing import Annotated  # pylint: disable=no-name-in-module
-else:
-    from typing_extensions import Annotated
-
+from typing_extensions import Annotated
 
 ColorMapName = Enum(  # type: ignore
     "ColorMapName", [(a, a) for a in sorted(cmap.list())]