+++ src/titiler/core/titiler/core/models/mapbox.py
@@ -1,16 +1,8 @@
 """Common response models."""
 
-from enum import Enum
-from typing import List, Optional, Tuple
+from typing import List, Literal, Optional, Tuple
 
-from pydantic import BaseModel, Field, root_validator
-
-
-class SchemeEnum(str, Enum):
-    """TileJSON scheme choice."""
-
-    xyz = "xyz"
-    tms = "tms"
+from pydantic import BaseModel, Field, model_validator
 
 
 class TileJSON(BaseModel):
@@ -22,34 +14,29 @@ class TileJSON(BaseModel):
     """
 
     tilejson: str = "2.2.0"
-    name: Optional[str]
-    description: Optional[str]
+    name: Optional[str] = None
+    description: Optional[str] = None
     version: str = "1.0.0"
-    attribution: Optional[str]
-    template: Optional[str]
-    legend: Optional[str]
-    scheme: SchemeEnum = SchemeEnum.xyz
+    attribution: Optional[str] = None
+    template: Optional[str] = None
+    legend: Optional[str] = None
+    scheme: Literal["xyz", "tms"] = "xyz"
     tiles: List[str]
-    grids: Optional[List[str]]
-    data: Optional[List[str]]
+    grids: Optional[List[str]] = None
+    data: Optional[List[str]] = None
     minzoom: int = Field(0, ge=0, le=30)
     maxzoom: int = Field(30, ge=0, le=30)
     bounds: List[float] = [-180, -90, 180, 90]
-    center: Optional[Tuple[float, float, int]]
+    center: Optional[Tuple[float, float, int]] = None
 
-    @root_validator
-    def compute_center(cls, values):
+    @model_validator(mode="after")
+    def compute_center(self):
         """Compute center if it does not exist."""
-        bounds = values["bounds"]
-        if not values.get("center"):
-            values["center"] = (
+        bounds = self.bounds
+        if not self.center:
+            self.center = (
                 (bounds[0] + bounds[2]) / 2,
                 (bounds[1] + bounds[3]) / 2,
-                values["minzoom"],
+                self.minzoom,
             )
-        return values
-
-    class Config:
-        """TileJSON model configuration."""
-
-        use_enum_values = True
+        return self